refactor(client): flatten entity factory package
