+++ client/src/main/java/net/lapidist/colony/client/entities/BuildingFactory.java
@@ -1,4 +1,4 @@
-package net.lapidist.colony.client.entities.factories;
+package net.lapidist.colony.client.entities;
 
 import com.artemis.Entity;
 import com.artemis.World;
@@ -8,7 +8,7 @@ import net.lapidist.colony.registry.BuildingDefinition;
 import net.lapidist.colony.registry.Registries;
 import java.util.Locale;
 
-import static net.lapidist.colony.client.entities.factories.SpriteFactoryUtil.createEntity;
+import static net.lapidist.colony.client.entities.SpriteFactoryUtil.createEntity;
 
 /**
  * Factory methods for creating building entities.
