--- client/src/main/java/net/lapidist/colony/client/entities/factories/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Factory classes for creating client-side entities.
- */
-package net.lapidist.colony.client.entities.factories;
