+++ tests/src/test/java/net/lapidist/colony/tests/client/entities/factories/TileFactoryTest.java
@@ -3,7 +3,7 @@ package net.lapidist.colony.tests.client.entities.factories;
 import com.artemis.World;
 import com.artemis.WorldConfigurationBuilder;
 import com.badlogic.gdx.math.Vector2;
-import net.lapidist.colony.client.entities.factories.TileFactory;
+import net.lapidist.colony.client.entities.TileFactory;
 import net.lapidist.colony.components.maps.TileComponent;
 import net.lapidist.colony.components.resources.ResourceComponent;
 import net.lapidist.colony.components.state.ResourceData;
